requirements.txt
Django==1.10
django-cors-headers==2.2.0
djangorestframework==3.8.2
Pillow==5.1.0
psycopg2==2.7.4
